Exit run() when CA certs change (#18)
Adds a watcher for the CA files specified in the log and metric exporter
configuration. When a change is detected to any of the CA files, the
run() method exits. Typically, the detected changes will be caused by
rotating CA certificates. The expectation is that the application will
restart and/or re-instantiate the `Otel` component when `run()` exits.

Integration tests have been updated to confirm `run()` exits when a CA
file changes. The changes have also been integrated and tested with AIO
Dataflows pods.